move docblock rename namespace to DocBlockNamespaceRenamer service
